Change data type of InvoiceValue to double
In the Intervention entity, the data type of the InvoiceValue property was initially a nullable decimal. This commit changes it to a nullable double to better align with the required precision of monetary values.

+++ ESIN.Clinic.Infrastructure/Entities/Intervention.cs
@@ -6,7 +6,7 @@ public class Intervention
     public DateTime ReportDate { get; set; }
     public string? Observations { get; set; }
     public DateTime? EvaluationDate { get; set; }
-    public decimal? InvoiceValue { get; set; }
+    public double? InvoiceValue { get; set; }
     public int InterventionType { get; set; }
     public DateTime? StartDate { get; set; }
     public DateTime? EndDate { get; set; }
